register trace clock freq from target DataTraceValue payload
Instead of a required --freq option, the frequency of the trace clock is
now always expected to be sent over ITM by the target via a valid
DataTraceValue payload. A valid payload is non-zero and 4 bytes long.
Only the first matching payload of this type is intercepted; subsuquent
payloads are forwarded as per usual. The payload is interpreted as
little endian.

A side-effect of this feature is that any ITM packets received before
this payload are dropped. In my local tests these packets have usually
been noise, so it's okay for now (an initial release that is), but we
may want to recover them later on.

Closes #4.

+++ src/recovery.rs
@@ -30,11 +30,11 @@ type SwAssocs = BTreeMap<SwExceptionNumber, Vec<String>>;
 pub struct Metadata {
     maps: TaskResolveMaps,
     timestamp: chrono::DateTime<Local>,
-    freq: usize,
+    freq: u32,
 }
 
 impl Metadata {
-    pub fn new(maps: TaskResolveMaps, timestamp: chrono::DateTime<Local>, freq: usize) -> Self {
+    pub fn new(maps: TaskResolveMaps, timestamp: chrono::DateTime<Local>, freq: u32) -> Self {
         Self {
             maps,
             timestamp,
